requirement.txt
# ==============================
# Web Framework
# ==============================
Flask==3.0.2
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0

# ==============================
# Data Processing
# ==============================
pandas==2.2.2
numpy==1.26.4

# ==============================
# Machine Learning
# ==============================
scikit-learn==1.4.2
sentence-transformers==2.7.0
torch==2.2.2

# ==============================
# Deep Learning (HR Model)
# ==============================
tensorflow==2.15.0
keras==2.15.0

# ==============================
# NLP
# ==============================
spacy==3.7.4

# ==============================
# Resume Parsing
# ==============================
pdfplumber==0.10.3
python-docx==1.1.0

# ==============================
# HR Video Analysis
# ==============================
opencv-python==4.9.0.80
mediapipe==0.10.11
